backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1

dash==2.17.0
dash-bootstrap-components==1.6.0
plotly==5.22.0

numpy==2.4.2
pandas==3.0.1
python-dateutil==2.9.0.post0
six==1.17.0
tzdata==2025.3
psycopg[binary]==3.3.3
transformers==4.54.0
torch==2.11.0
scikit-learn==1.8.0
